fix race conditions when pausing the consumer
When calling pause() the calling thread will now block until all handlers have processed their message.
In addition, messages that have been prefetched by the client library will now be ignored if there are no more channels open.
The latter race condition is harmless, as it would've only logged exceptions, nevertheless it might be confused with an actual error so it's better to handle it explicitely.

 renamed process() to doProcess() in MessageHandler to wrap the RW locking