speed up test suite
On some Macs, the full test suite takes more than 15 minutes to run.
This is caused by using "localhost" when connecting to docker containers.
Relacing getContainerIpAdress() with "127.0.0.1" fixes the problem.
Apparently this has something to do with IPV6 where localhost is
resolved to ::1 and trying to connect to ::1:port leads to a timeout
when the process listening on that address is a docker container.